requirements.txt
Pillow
scikit-image
opencv-python
numpy
imageio
scipy==1.2.0
